changed remaining occurances of "unsigned long" into uint32_t
    - makes sure integers have save width across platforms
    - addresses issue eudaq/eudaq#61

+++ main/include/eudaq/Event.hh
@@ -108,7 +108,7 @@ namespace eudaq {
       static event_creator GetCreator(uint32_t id);
 
     private:
-      typedef std::map<unsigned long, event_creator> map_t;
+      typedef std::map<uint32_t, event_creator> map_t;
       static map_t & get_map();
   };
